Extract isValueAllowed into shared pkg/validator.IsValueAllowed helper
