+++ pkg/validator/github/validator_test.go
@@ -151,67 +151,3 @@ func TestCheckAllowLists(t *testing.T) {
 	}
 }
 
-func TestIsValueAllowed(t *testing.T) {
-	tests := []struct {
-		name     string
-		value    string
-		allowed  []string
-		expected bool
-	}{
-		{
-			name:     "exact_match",
-			value:    "my-org/my-repo",
-			allowed:  []string{"my-org/my-repo"},
-			expected: true,
-		},
-		{
-			name:     "no_match",
-			value:    "my-org/my-repo",
-			allowed:  []string{"other-org/other-repo"},
-			expected: false,
-		},
-		{
-			name:     "wildcard_suffix",
-			value:    "my-org/my-repo",
-			allowed:  []string{"my-org/*"},
-			expected: true,
-		},
-		{
-			name:     "wildcard_no_match",
-			value:    "other-org/my-repo",
-			allowed:  []string{"my-org/*"},
-			expected: false,
-		},
-		{
-			name:     "multiple_allowed",
-			value:    "org-b/repo",
-			allowed:  []string{"org-a/*", "org-b/*"},
-			expected: true,
-		},
-		{
-			name:     "empty_value",
-			value:    "",
-			allowed:  []string{"my-org/*"},
-			expected: false,
-		},
-		{
-			name:     "empty_allowed_list",
-			value:    "my-org/my-repo",
-			allowed:  []string{},
-			expected: false,
-		},
-		{
-			name:     "star_only_wildcard",
-			value:    "anything",
-			allowed:  []string{"*"},
-			expected: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := isValueAllowed(tt.value, tt.allowed)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
